+++ packages/react/src/components/design-wrapper.tsx
@@ -12,21 +12,25 @@ type DesignWrapperProps = {
 };
 
 export const DesignWrapper = (props: DesignWrapperProps) => {
-  const { children, className, styles } = props;
+  const { children, className } = props;
 
-  const { container = true, background, ...rest } = styles || {};
+  const styles = props.styles?.desktop ?? {};
+
+  const { background } = styles;
 
   const matches =
     typeof background === "string"
       ? background.match(/url\(["']?(.*?)["']?\)/)
       : undefined;
+
   const src = matches ? matches[1] : undefined;
 
+  const container = props.styles?.container ?? true;
+
+  const id = React.useId().replace(/[^a-zA-Z0-9]/g, "");
+
   return (
-    <div
-      className={cn("sb-relative", className)}
-      style={{ ...(!src && { background }), ...rest }}
-    >
+    <div id={id} className={cn("sb-relative", className)}>
       {src && (
         <Image
           src={src}
@@ -40,6 +44,26 @@ export const DesignWrapper = (props: DesignWrapperProps) => {
       ) : (
         children
       )}
+
+      <style>
+        {`
+        #${id} {
+          ${Object.entries(props.styles?.desktop ?? {})
+            .map(([key, value]) => `${key}: ${value};`)
+            .join(" ")}
+        }
+
+        @media (max-width: 1024px) {
+          #${id} {
+            ${Object.entries(props.styles?.mobile ?? {})
+              .map(([key, value]) => `${key}: ${value}; `)
+              .join(" ")}
+          }
+        }
+      `.replace(/\s\s+/g, " ")}
+      </style>
     </div>
   );
 };
+
+export default DesignWrapper;