+++ packages/react/tsup.config.ts
@@ -1,7 +1,6 @@
+import * as fs from "node:fs/promises";
+import * as path from "node:path";
 import { defineConfig, Options } from "tsup";
-
-import * as fs from 'node:fs/promises';
-import * as path from 'node:path';
 
 const cfg: Options = {
   splitting: true, //error triggerer
@@ -21,16 +20,19 @@ const config = defineConfig([
     outDir: "dist",
     onSuccess: async () => {
       const files = await fs.readdir("./dist");
-      
+
       for (const file of files) {
-        if (file.startsWith("build-container-client") && (file.endsWith(".mjs") || file.endsWith(".js"))) {
+        if (
+          file.startsWith("build-container-client") &&
+          (file.endsWith(".mjs") || file.endsWith(".js"))
+        ) {
           const filePath = path.join("./dist", file);
-          const data = await fs.readFile(filePath, 'utf8');
+          const data = await fs.readFile(filePath, "utf8");
           const updatedContent = `'use client';\n${data}`;
-          await fs.writeFile(filePath, updatedContent, 'utf8');
+          await fs.writeFile(filePath, updatedContent, "utf8");
         }
       }
-    }
+    },
   },
   // {
   //   ...cfg,