+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @simple-builder/react
 
+## 1.0.6
+
+### Patch Changes
+
+- Replaced cva with tailwind-variants, added new tabs variant, mobile en desktop design support.
+- Updated dependencies
+  - @simple-builder/server@1.0.6
+
 ## 1.0.5
 
 ### Patch Changes